requirements.txt
minio
python-dotenv
pymongo
requests

